MultiAZAutoScalingWithELBInVPC
Sample Template: Create a multi-az, load-balanced and Auto Scaled sample website running on an Apache Web Server. The application is configured to span all Availability Zones in the region and is Auto-Scaled based on the CPU utilization of the web servers. Notifications will be sent to the operator's email address on scaling events. The instances are load-balanced with a simple health check against the default web page.